Fibonacci test apparently going swimmingly
"asmtosia.py" is added the asembler code itself, can be modified
to have numbers other than 8, 9, etc. "fib.*" files for sanity
check, output is compared with Vivado's fibonacci waveform
(to-be-added in later commit). "imem.{dat;dat.verbose}" now have
the machine code for the Fibonacci stuff and the ".asm" file has the
codez for Fibonacci as well.